Avoid integer overflow

+++ src/nf.c
@@ -96,7 +96,7 @@ main(int argc, char *argv[])
         exit(EXIT_FAILURE);
     }
 
-    init_shm(&shm, shm.head, BUF_NUM * MDBUF_SIZE, BUF_NUM * MBUF_PKTBUF_SIZE, sizeof(struct desc) * opt.vq_size);
+    init_shm(&shm, shm.head, (size_t)BUF_NUM * (size_t)MDBUF_SIZE, (size_t)BUF_NUM * (size_t)MBUF_PKTBUF_SIZE, sizeof(struct desc) * opt.vq_size);
 
     memset(shm.head, 0, SHM_SIZE);
