+++ src/tx.c
@@ -41,7 +41,7 @@ main(int argc, char *argv[])
         perror("mmap");
         exit(EXIT_FAILURE);
     }
-    init_shm(&shm, shm.head, BUF_NUM * MDBUF_SIZE, BUF_NUM * MBUF_PKTBUF_SIZE, sizeof(struct desc) * opt.vq_size);
+    init_shm(&shm, shm.head, (size_t)BUF_NUM * (size_t)MDBUF_SIZE, (size_t)BUF_NUM * (size_t)MBUF_PKTBUF_SIZE, sizeof(struct desc) * opt.vq_size);
 
     init_mpools(&mpools_host, MDBUF_SIZE, MBUF_PKTBUF_SIZE, BUF_NUM, opt.mobj_cache_num,
         shm.head + shm.end_offset + CACHE_LINE_SIZE + BUF_NUM * (MDBUF_SIZE + MBUF_PKTBUF_SIZE), NULL);
